Added the project 'ChessVideoAnalyser'
This commit starts the project. The code uses pafy to download the video
https://www.youtube.com/watch?v=7bubb3Wn-48, opens it with imageio and
cuts out a square with position (170:248, 61:139), corresponding to the
board in the bottom right of the video. This cutout is then saved for
each second in 'images/board#s.jpg' where # is the current second.

to run, first install imageio and pafy with 'pip install pafy' and
'pip install imageio'